post-review fixes for wildcard internal transitions support
